refactor: extract useRecentGraphs composable (recentOpen, refreshRecent, onDeleteRecent, endpointKey)
